backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
websocket-client>=1.7.0
pymongo>=4.6.0
PyJWT>=2.8.0
bcrypt>=4.0.0
requests>=2.28.0
python-docx>=1.0.0
